Detailed comments, intuitive organization
