+++ main.go
@@ -19,6 +19,7 @@ func main() {
 	// Set up routes
 	routes.SetupUserRoutes()
 	routes.SetupAuthRoutes()
+	routes.SetupRecipeRoutes()
 
 	fmt.Println("Server is running on http://localhost:8080")
 	log.Fatal(http.ListenAndServe(":8080", nil))
